+++ spec/spec_helper.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'bundler/setup'
-require 'omniauth-line-v2_1'
+require 'omniauth-line'
 require 'webmock/rspec'
 
 RSpec.configure do |config|