+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ## [1.1.0] - 2025-08-02
 
-- Refactor: Refactor `OmniAuth::Strategies::LineV21#authorize_params`
+- Refactor: Refactor `OmniAuth::Strategies::Line#authorize_params`
 - Test: Update test
 
 ## [1.0.0] - 2025-08-01